[Root6] Implement custom TRint wrapper
Provide a simple copypasta of ROOT's main program for an interactive
session using a TRint Application. Prepare gEnv and gROOT before
starting application so that we can experiment with using custom
rc files, logon/off scripts and macro/loader paths. This functionality
may be useful for interpreters for projects with specific requirements
(e.g. libs/dicts of data types custom styles etc). It's possible
ROOT's own config system may be usable for this, but it's singular
so that different configs are difficult to manage.

Use Binreloc for application self location.