Minimal test of storing variant types
Cannot yet store variant types in ROOT file directly. Experiment
with a workaround based on using Boost's variant type as a transient
data member of a class with a custom Streamer method. Use type
switching to store and retrieve the type index of the variant,
using this information to subsequently store/retrieve the value.

Add a more advanced Dynamic type to explore user interface to a
variant based object.

Simplify and remove some earlier dictionary generation for Boost/Poco
types as these seem not to work fully due to unknown types or lack
of default constructors.